+++ packages/api-rest/src/apis/common/internalPost.ts
@@ -6,6 +6,7 @@ import { AmplifyClassV6 } from '@aws-amplify/core';
 import { InternalPostInput, RestApiResponse } from '../../types';
 import { createCancellableOperation } from '../../utils';
 import { CanceledError } from '../../errors';
+import { isIamAuthApplicableForGraphQL } from '../../utils/isIamAuthApplicable';
 
 import { transferHandler } from './handler';
 
@@ -46,7 +47,7 @@ const cancelTokenMap = new WeakMap<Promise<any>, AbortController>();
  * @param postInput.abortController The abort controller used to cancel the POST request
  * @returns a {@link RestApiResponse}
  *
- * @throws an {@link Error} with `Network error` as the `message` when the external resource is unreachable due to one
+ * @throws an {@link AmplifyError} with `Network Error` as the `message` when the external resource is unreachable due to one
  * of the following reasons:
  *   1. no network connection
  *   2. CORS error
@@ -66,6 +67,7 @@ export const post = (
 				...options,
 				abortSignal: controller.signal,
 			},
+			isIamAuthApplicableForGraphQL,
 			options?.signingServiceInfo,
 		);