+++ packages/auth/src/providers/cognito/apis/updateUserAttributes.ts
@@ -16,13 +16,14 @@ import {
 	UpdateUserAttributesInput,
 	UpdateUserAttributesOutput,
 } from '../types';
-import { updateUserAttributes as updateUserAttributesClient } from '../utils/clients/CognitoIdentityProvider';
 import { assertAuthTokens } from '../utils/types';
-import { getRegion } from '../utils/clients/CognitoIdentityProvider/utils';
+import { getRegionFromUserPoolId } from '../../../foundation/parsers';
 import { toAttributeType } from '../utils/apiHelpers';
-import { CodeDeliveryDetailsType } from '../utils/clients/CognitoIdentityProvider/types';
+import { CodeDeliveryDetailsType } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider/types';
 import { UpdateUserAttributesException } from '../types/errors';
 import { getAuthUserAgentValue } from '../../../utils';
+import { createUpdateUserAttributesClient } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../factories';
 
 /**
  * Updates user's attributes while authenticated.
@@ -39,11 +40,17 @@ export const updateUserAttributes = async (
 	const authConfig = Amplify.getConfig().Auth?.Cognito;
 	const clientMetadata = options?.clientMetadata;
 	assertTokenProviderConfig(authConfig);
+	const { userPoolEndpoint, userPoolId } = authConfig;
 	const { tokens } = await fetchAuthSession({ forceRefresh: false });
 	assertAuthTokens(tokens);
+	const updateUserAttributesClient = createUpdateUserAttributesClient({
+		endpointResolver: createCognitoUserPoolEndpointResolver({
+			endpointOverride: userPoolEndpoint,
+		}),
+	});
 	const { CodeDeliveryDetailsList } = await updateUserAttributesClient(
 		{
-			region: getRegion(authConfig.userPoolId),
+			region: getRegionFromUserPoolId(userPoolId),
 			userAgentValue: getAuthUserAgentValue(AuthAction.UpdateUserAttributes),
 		},
 		{