+++ packages/storage/src/providers/s3/apis/internal/copy.ts
@@ -21,6 +21,8 @@ import { assertValidationError } from '../../../../errors/utils/assertValidation
 import { copyObject } from '../../utils/client/s3data';
 import { getStorageUserAgentValue } from '../../utils/userAgent';
 import { logger } from '../../../../utils';
+// TODO: Remove this interface when we move to public advanced APIs.
+import { CopyInput as CopyWithPathInputWithAdvancedOptions } from '../../../../internals';
 
 const isCopyInputWithPath = (
 	input: CopyInput | CopyWithPathInput,
@@ -44,7 +46,7 @@ const storageBucketAssertion = (
 
 export const copy = async (
 	amplify: AmplifyClassV6,
-	input: CopyInput | CopyWithPathInput,
+	input: CopyInput | CopyWithPathInputWithAdvancedOptions,
 ): Promise<CopyOutput | CopyWithPathOutput> => {
 	return isCopyInputWithPath(input)
 		? copyWithPath(amplify, input)
@@ -53,7 +55,7 @@ export const copy = async (
 
 const copyWithPath = async (
 	amplify: AmplifyClassV6,
-	input: CopyWithPathInput,
+	input: CopyWithPathInputWithAdvancedOptions,
 ): Promise<CopyWithPathOutput> => {
 	const { source, destination } = input;