+++ packages/storage/src/providers/s3/apis/internal/getProperties.ts
@@ -7,7 +7,6 @@ import { StorageAction } from '@aws-amplify/core/internals/utils';
 import {
 	GetPropertiesInput,
 	GetPropertiesOutput,
-	GetPropertiesWithPathInput,
 	GetPropertiesWithPathOutput,
 } from '../../types';
 import {
@@ -18,10 +17,12 @@ import { headObject } from '../../utils/client/s3data';
 import { getStorageUserAgentValue } from '../../utils/userAgent';
 import { logger } from '../../../../utils';
 import { STORAGE_INPUT_KEY } from '../../utils/constants';
+// TODO: Remove this interface when we move to public advanced APIs.
+import { GetPropertiesInput as GetPropertiesWithPathInputWithAdvancedOptions } from '../../../../internals';
 
 export const getProperties = async (
 	amplify: AmplifyClassV6,
-	input: GetPropertiesInput | GetPropertiesWithPathInput,
+	input: GetPropertiesInput | GetPropertiesWithPathInputWithAdvancedOptions,
 	action?: StorageAction,
 ): Promise<GetPropertiesOutput | GetPropertiesWithPathOutput> => {
 	const { s3Config, bucket, keyPrefix, identityId } =