+++ packages/storage/src/providers/s3/apis/internal/remove.ts
@@ -4,12 +4,7 @@
 import { AmplifyClassV6 } from '@aws-amplify/core';
 import { StorageAction } from '@aws-amplify/core/internals/utils';
 
-import {
-	RemoveInput,
-	RemoveOutput,
-	RemoveWithPathInput,
-	RemoveWithPathOutput,
-} from '../../types';
+import { RemoveInput, RemoveOutput, RemoveWithPathOutput } from '../../types';
 import {
 	resolveS3ConfigAndInput,
 	validateStorageOperationInput,
@@ -18,10 +13,12 @@ import { deleteObject } from '../../utils/client/s3data';
 import { getStorageUserAgentValue } from '../../utils/userAgent';
 import { logger } from '../../../../utils';
 import { STORAGE_INPUT_KEY } from '../../utils/constants';
+// TODO: Remove this interface when we move to public advanced APIs.
+import { RemoveInput as RemoveWithPathInputWithAdvancedOptions } from '../../../../internals';
 
 export const remove = async (
 	amplify: AmplifyClassV6,
-	input: RemoveInput | RemoveWithPathInput,
+	input: RemoveInput | RemoveWithPathInputWithAdvancedOptions,
 ): Promise<RemoveOutput | RemoveWithPathOutput> => {
 	const { s3Config, keyPrefix, bucket, identityId } =
 		await resolveS3ConfigAndInput(amplify, input);