+++ packages/storage/src/providers/s3/types/options.ts
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { StorageAccessLevel } from '@aws-amplify/core';
-import { AWSCredentials } from '@aws-amplify/core/internals/utils';
 import {
 	CredentialsProviderOptions,
 	SigningOptions,
 } from '@aws-amplify/core/internals/aws-client-utils';
+import { AWSCredentials } from '@aws-amplify/core/internals/utils';
 
 import { TransferProgressEvent } from '../../../types';
 import {
@@ -26,6 +26,11 @@ export type AWSTemporaryCredentials = Required<
 >;
 
 /**
+ * Async function returning AWS credentials for an API call. This function
+ * is invoked with S3 locations(bucket and path).
+ * If omitted, the global credentials configured in Amplify Auth
+ * would be used.
+ *
  * @internal
  */
 export type LocationCredentialsProvider = (
@@ -45,13 +50,6 @@ interface CommonOptions {
 	 */
 	useAccelerateEndpoint?: boolean;
 
-	/**
-	 * Async function returning AWS credentials for an API call. This function
-	 * is invoked with S3 locations(bucket and path).
-	 * If omitted, the global credentials configured in Amplify Auth
-	 * would be used.
-	 */
-	locationCredentialsProvider?: LocationCredentialsProvider;
 	bucket?: StorageBucket;
 }
 
@@ -103,9 +101,9 @@ interface TransferOptions {
 /**
  * Input options type for S3 getProperties API.
  */
-/** @deprecated Use {@link GetPropertiesOptionsWithPath} instead. */
-export type GetPropertiesOptionsWithKey = ReadOptions & CommonOptions;
-export type GetPropertiesOptionsWithPath = CommonOptions;
+/** @deprecated Use {@link GetPropertiesWithPathOptions} instead. */
+export type GetPropertiesWithKeyOptions = ReadOptions & CommonOptions;
+export type GetPropertiesWithPathOptions = CommonOptions;
 
 /**
  * Input options type for S3 getProperties API.
@@ -113,25 +111,25 @@ export type GetPropertiesOptionsWithPath = CommonOptions;
 export type RemoveOptions = WriteOptions & CommonOptions;
 
 /**
- * @deprecated Use {@link ListAllOptionsWithPath} instead.
+ * @deprecated Use {@link ListAllWithPathOptions} instead.
  * Input options type with prefix for S3 list all API.
  */
-export type ListAllOptionsWithPrefix = StorageListAllOptions &
+export type ListAllWithPrefixOptions = StorageListAllOptions &
 	ReadOptions &
 	CommonOptions;
 
 /**
- * @deprecated Use {@link ListPaginateOptionsWithPath} instead.
+ * @deprecated Use {@link ListPaginateWithPathOptions} instead.
  * Input options type with prefix for S3 list API to paginate items.
  */
-export type ListPaginateOptionsWithPrefix = StorageListPaginateOptions &
+export type ListPaginateWithPrefixOptions = StorageListPaginateOptions &
 	ReadOptions &
 	CommonOptions;
 
 /**
  * Input options type with path for S3 list all API.
  */
-export type ListAllOptionsWithPath = Omit<
+export type ListAllWithPathOptions = Omit<
 	StorageListAllOptions,
 	'accessLevel'
 > &
@@ -142,7 +140,7 @@ export type ListAllOptionsWithPath = Omit<
 /**
  * Input options type with path for S3 list API to paginate items.
  */
-export type ListPaginateOptionsWithPath = Omit<
+export type ListPaginateWithPathOptions = Omit<
 	StorageListPaginateOptions,
 	'accessLevel'
 > &
@@ -179,9 +177,9 @@ export type GetUrlOptions = CommonOptions & {
 	contentType?: string;
 };
 
-/** @deprecated Use {@link GetUrlOptionsWithPath} instead. */
-export type GetUrlOptionsWithKey = ReadOptions & GetUrlOptions;
-export type GetUrlOptionsWithPath = GetUrlOptions;
+/** @deprecated Use {@link GetUrlWithPathOptions} instead. */
+export type GetUrlWithKeyOptions = ReadOptions & GetUrlOptions;
+export type GetUrlWithPathOptions = GetUrlOptions;
 
 /**
  * Input options type for S3 downloadData API.
@@ -190,9 +188,9 @@ export type DownloadDataOptions = CommonOptions &
 	TransferOptions &
 	BytesRangeOptions;
 
-/** @deprecated Use {@link DownloadDataOptionsWithPath} instead. */
-export type DownloadDataOptionsWithKey = ReadOptions & DownloadDataOptions;
-export type DownloadDataOptionsWithPath = DownloadDataOptions;
+/** @deprecated Use {@link DownloadDataWithPathOptions} instead. */
+export type DownloadDataWithKeyOptions = ReadOptions & DownloadDataOptions;
+export type DownloadDataWithPathOptions = DownloadDataOptions;
 
 export type UploadDataChecksumAlgorithm = 'crc-32';
 
@@ -234,12 +232,12 @@ export type UploadDataOptions = CommonOptions &
 		checksumAlgorithm?: UploadDataChecksumAlgorithm;
 	};
 
-/** @deprecated Use {@link UploadDataOptionsWithPath} instead. */
-export type UploadDataOptionsWithKey = WriteOptions & UploadDataOptions;
-export type UploadDataOptionsWithPath = UploadDataOptions;
+/** @deprecated Use {@link UploadDataWithPathOptions} instead. */
+export type UploadDataWithKeyOptions = WriteOptions & UploadDataOptions;
+export type UploadDataWithPathOptions = UploadDataOptions;
 
 /** @deprecated This may be removed in the next major version. */
-export type CopySourceOptionsWithKey = ReadOptions & {
+export type CopySourceWithKeyOptions = ReadOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
 	bucket?: StorageBucket;
@@ -248,7 +246,7 @@ export type CopySourceOptionsWithKey = ReadOptions & {
 };
 
 /** @deprecated This may be removed in the next major version. */
-export type CopyDestinationOptionsWithKey = WriteOptions & {
+export type CopyDestinationWithKeyOptions = WriteOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
 	bucket?: StorageBucket;